Enable large tensor support for batch flatten

+++ src/operator/tensor/matrix_op-inl.h
@@ -243,7 +243,7 @@ inline bool FlattenShape(const nnvm::NodeAttrs& attrs,
   CHECK_EQ(out_attrs->size(), 1U);
   const mxnet::TShape &dshape = (*in_attrs)[0];
   if (!shape_is_known(dshape)) return false;
-  int target_dim = 1;
+  size_t target_dim = 1;
   for (int i = 1; i < dshape.ndim(); ++i) {
     target_dim *= dshape[i];
   }
